+++ rust/stackable-cockpit/src/platform/credentials.rs
@@ -11,7 +11,10 @@ pub type Result<T, E = Error> = std::result::Result<T, E>;
 #[derive(Debug, Snafu)]
 pub enum Error {
     #[snafu(display("failed to fetch data from Kubernetes API"))]
-    KubeClientFetch { source: k8s::Error },
+    KubeClientFetch {
+        #[snafu(source(from(k8s::Error, Box::new)))]
+        source: Box<k8s::Error>,
+    },
 
     #[snafu(display("no credentials secret found"))]
     NoSecret,