+++ rust/xtask/src/openapi.rs
@@ -35,9 +35,12 @@ pub fn generate() -> Result<(), GenOpenapiError> {
         .write_all(openapi_json.as_bytes())
         .context(WriteOpenapiSchemaSnafu)?;
     let status = codegen.wait().context(ImportOpenapiSchemaRunSnafu)?;
-    ensure!(status.success(), ImportOpenapiSchemaSnafu {
-        error_code: status.code()
-    });
+    ensure!(
+        status.success(),
+        ImportOpenapiSchemaSnafu {
+            error_code: status.code()
+        }
+    );
 
     Ok(())
 }